Relocate `/profile lastfm` to `/lastfm profile`
